flight-software/payload/jetson/requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.7
action-tutorials-py==0.20.7
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
builtin-interfaces==1.2.2
composition-interfaces==1.2.2
control-msgs==4.8.0
controller-manager==2.53.0
controller-manager-msgs==2.53.0
cv-bridge==3.2.1
demo-nodes-py==0.20.7
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
eclipse-zenoh==1.7.2
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
generate-parameter-library-py==0.6.0
geometry-msgs==4.9.0
image-geometry==3.2.1
interactive-markers==2.3.2
laser-geometry==2.4.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
lifecycle-msgs==1.2.2
logging-demo==0.20.7
map-msgs==2.1.0
message-filters==4.3.11
mocap4r2-control-msgs==0.0.7
moveit-msgs==2.2.1
nav-msgs==4.9.0
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
osrf-pycommon==2.1.6
pcl-msgs==1.0.0
pendulum-msgs==0.20.7
play-motion-builder-msgs==1.4.0
play-motion2==1.8.2
play-motion2-msgs==1.8.2
python-qt-binding==1.1.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.7
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.8
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2plugin==5.1.3
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rpyutils==0.2.2
rqt==1.1.9
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-controller-manager==2.53.0
rqt-dotgraph==0.0.5
rqt-gauges==0.0.3
rqt-graph==1.3.1
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-joint-trajectory-controller==2.52.0
rqt-moveit==1.0.1
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.3
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.5
rqt-topic==1.5.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
srdfdom==2.0.7
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.25.18
tf2-kdl==0.25.18
tf2-msgs==0.25.18
tf2-py==0.25.18
tf2-ros-py==0.25.18
tf2-sensor-msgs==0.25.18
tf2-tools==0.25.18
topic-monitor==0.20.7
trajectory-msgs==4.9.0
turtlesim==1.4.3
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
visualization-msgs==4.9.0
